NoAdsHolder, date, decimal

+++ app/src/main/java/ragalik/baraxolka/other_logic/edit_creator/EditCreator.java
@@ -1,7 +1,6 @@
 package ragalik.baraxolka.other_logic.edit_creator;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatAutoCompleteTextView;
 import androidx.appcompat.widget.AppCompatButton;
@@ -13,39 +12,20 @@ import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bumptech.glide.Glide;
-import com.bumptech.glide.RequestBuilder;
-import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.request.FutureTarget;
-import com.bumptech.glide.request.Request;
-import com.bumptech.glide.request.target.SimpleTarget;
-import com.bumptech.glide.request.target.SizeReadyCallback;
-import com.bumptech.glide.request.target.Target;
-import com.bumptech.glide.request.target.ViewTarget;
-import com.bumptech.glide.request.transition.Transition;
-import com.bumptech.glide.signature.ObjectKey;
 import com.google.android.material.textfield.TextInputLayout;
-import com.squareup.picasso.MemoryPolicy;
-import com.squareup.picasso.NetworkPolicy;
-import com.squareup.picasso.Picasso;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.OutputStream;
-import java.net.URI;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -53,8 +33,6 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.TimeZone;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
@@ -66,8 +44,6 @@ import ragalik.baraxolka.network.entities.CategoryResponse;
 import ragalik.baraxolka.network.entities.ServerResponse;
 import ragalik.baraxolka.network.entities.Subcategories;
 import ragalik.baraxolka.other_logic.account.PathUtils;
-import ragalik.baraxolka.other_logic.ad_creator.AdCreator;
-import ragalik.baraxolka.other_logic.ad_creator.AdCreatorImageMenu;
 import ragalik.baraxolka.paging_feed.Ad;
 import ragalik.baraxolka.paging_feed.FullAdResponse;
 import retrofit2.Call;
@@ -575,7 +551,7 @@ public class EditCreator extends AppCompatActivity implements View.OnClickListen
             }
         }
 
-        int price = Integer.parseInt(priceEditText.getEditText().getText().toString());
+        double price = Double.parseDouble(priceEditText.getEditText().getText().toString());
         String datetime = dateFormat.format(new Date());
         String subcategory = subcategoryFromSpinner;
